fix: pamphlet party_size 컬럼 타입 불일치 수정
- pamphlets.party_size는 SMALLINT(스키마 기준)인데 엔티티가 Integer로 매핑돼 Hibernate validate 실패
- Pamphlet 엔티티 필드를 Short로 맞추고 요청/응답 DTO(Integer)와의 변환은 서비스 계층에서 처리

+++ src/main/java/com/mulmeong/domain/pamphlet/entity/Pamphlet.java
@@ -10,11 +10,11 @@ public class Pamphlet extends BaseTimeEntity {
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY) private Long id;
     @Column(name = "user_id", nullable = false) private Long userId;
     @Column(nullable = false) private String title;
-    @Column(name = "party_size") private Integer partySize;
+    @Column(name = "party_size") private Short partySize;
     @Column(name = "travel_date") private LocalDate travelDate;
     @Column(name = "cover_image_url") private String coverImageUrl;
     @Column(name = "share_token", unique = true) private String shareToken;
-    public Pamphlet(Long userId, String title, Integer partySize, LocalDate travelDate, String coverImageUrl, String shareToken) {
+    public Pamphlet(Long userId, String title, Short partySize, LocalDate travelDate, String coverImageUrl, String shareToken) {
         this.userId = userId; this.title = title; this.partySize = partySize; this.travelDate = travelDate;
         this.coverImageUrl = coverImageUrl; this.shareToken = shareToken;
     }
